+++ HISTORY.rst
@@ -4,6 +4,10 @@ Changelog
 0.0.2 (unreleased)
 ~~~~~~~~~~~~~~~~~~
 
+- Fix being unable to delete an allocation with a quota > 1. 
+  See issue #8.
+  [href]
+
 - Replace read session write guard with a simpler version.
   [href]